Moving all crypto relevant code to utility/crypto folder
